Plan 9 from Bell Labs 1990-10-12

+++ gnot/mem.h
@@ -21,8 +21,8 @@
 #define	HZ		(60)			/* clock frequency */
 #define	MS2HZ		(1000/HZ)		/* millisec per clock tick */
 #define	TK2SEC(t)	((t)/HZ)		/* ticks to seconds */
-#define	TK2MS(t)	(((t)*1000)/HZ)		/* ticks to milliseconds */
-#define	MS2TK(t)	(((t)*HZ)/1000)		/* milliseconds to ticks */
+#define	TK2MS(t)	((((ulong)(t))*1000)/HZ)	/* ticks to milliseconds */
+#define	MS2TK(t)	((((ulong)(t))*HZ)/1000)	/* milliseconds to ticks */
 
 /*
  * SR bits
